+++ flutter_secure_storage_platform_interface/lib/flutter_secure_storage_platform_interface.dart
@@ -1,5 +1,7 @@
 library flutter_secure_storage_platform_interface;
 
+import 'dart:io';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';