+++ flutter_secure_storage_platform_interface/lib/src/method_channel_flutter_secure_storage.dart
@@ -3,7 +3,23 @@ part of '../flutter_secure_storage_platform_interface.dart';
 const MethodChannel _channel =
     MethodChannel('plugins.it_nomads.com/flutter_secure_storage');
 
+const EventChannel _eventChannel =
+    EventChannel('plugins.it_nomads.com/flutter_secure_storage/events');
+
 class MethodChannelFlutterSecureStorage extends FlutterSecureStoragePlatform {
+  Stream<bool> get onCupertinoProtectedDataAvailabilityChanged => _eventChannel
+      .receiveBroadcastStream()
+      .where((event) => event is bool)
+      .map((event) => event as bool);
+
+  Future<bool?> isCupertinoProtectedDataAvailable() async {
+    if (kIsWeb || !(Platform.isIOS || Platform.isMacOS)) {
+      return null;
+    }
+    return (await _channel.invokeMethod<bool>('isProtectedDataAvailable')) ??
+        false;
+  }
+
   @override
   Future<bool> containsKey({
     required String key,