+++ flutter_secure_storage_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.2
+Adds onCupertinoProtectedDataAvailabilityChanged and isCupertinoProtectedDataAvailable via MethodChannelFlutterSecureStorage to prevent breaking changes.
+
+## 1.1.1
+Reverts onCupertinoProtectedDataAvailabilityChanged and isCupertinoProtectedDataAvailable.
+
+## 1.1.0
+Adds onCupertinoProtectedDataAvailabilityChanged and isCupertinoProtectedDataAvailable.
+
 ## 1.0.2
 - Update Dart SDK Constraint to support <4.0.0 instead of <3.0.0.